requirements.txt
anyio
certifi
click
colorama
dnspython
email-validator
fastapi
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
MarkupSafe
orjson
psycopg2
pydantic
python-dotenv
python-multipart
PyYAML
rfc3986
six
sniffio
SQLAlchemy
starlette
typing_extensions
ujson
uvicorn
watchfiles
websockets
